Add files via upload
Changed the way I was storing table metadata, now there is metadata on field names and primary key which are in its own table in the database.  This way they are persistent.  If there is a table from a SELECT statement returned, that temporary table is not in the database so its TABLE_NAME is attached as a property.  Likewise with anything else that is not persistent in the database.  Also added SHOW TABLES and DESCRIBE and now have a run box where SQL statements can be executed and a div to see the results.  The results currently are in JSON so the next thing I would like to do is format them.  So it may take a while to come up with that algorithm.  It needs to be able to plot stuff with missing data etc.